Reorganize mipego
