+++ examples/sin/mae-rand-samp-sin.json
@@ -13,7 +13,7 @@
 "truncated_upper": 2.0,
 "threshold": 0.01,
 "nn_builder_class": "dlopt.nn.RNNBuilder",
-"dense_activation": "tanh",
+"dense_activation": "sigmoid",
 "output_logger_class": "dlopt.util.JSONOutput",
 "output_logger_params": {"filename": "output.json"},
 "data_loader_class": "examples.sin.sin.SinDataLoader",
