+++ publications/mipego/.gitignore
